RumBoardProject/src/backend/requirements.txt
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.3
cryptography==2.8
distlib==0.3.4
distro==1.4.0
docopt==0.6.2
entrypoints==0.3
fasteners==0.14.1
filelock==3.6.0
Flask==2.0.3
Flask-Cors==3.0.10
future==0.18.2
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
itsdangerous==2.1.1
Jinja2==3.0.3
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pbr==5.8.1
pexpect==4.6.0
Pillow==9.1.0
pipenv==2022.3.24
pipreqs==0.4.11
platformdirs==2.5.1
protobuf==3.6.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycparser==2.21
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-dateutil==2.7.3
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
testresources==2.0.1
urllib3==1.25.8
virtualenv==20.14.0
virtualenv-clone==0.5.7
wadllib==1.3.3
Werkzeug==2.0.3
yarg==0.1.9
